+++ tsdown.config.ts
@@ -1,8 +1,17 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-	entry: ['src/index.ts'],
-	format: ['esm', 'cjs'],
+	entry: 'src/index.ts',
 	dts: true,
+	exports: true,
 	clean: true,
+	target: 'esnext',
+	unused: true,
+	onSuccess: 'bun fmt package.json',
+	unbundle: true,
+	minify: 'dce-only',
+	publint: true,
+	attw: {
+		ignoreRules: ['cjs-resolves-to-esm', 'no-resolution'],
+	},
 });